Minor tweaks after reading the code
The sets functionality is now a natively implemented in Python, so there
is no need to import it. In fact, it gives an error because it doesn't
exists anymore.
Also, I incremented the number of generation to 100.
Good (or bad) news, with the map of Europe and the actual code, we got
down to 16 colors. We have a long way to go :+1: