Initial population generation function implemented.
The config file was updated to run with the "heuristic generation of the
population", this can be changed to "random" to run the original
function.
Notes: This new population might be better for the TTP but not for the
TSP and evolution is killing most of them.
For this to work a fitness function considering the TTP must be
implemented as well.